+++ packages/web-forms/src/components/common/CheckboxWidget.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
-import TextMedia from '@/components/common/TextMedia.vue';
-import { selectOptionId } from '@/lib/format/ids.ts';
+import TextMedia from '@getodk/web-forms/components/common/TextMedia.vue';
+import { selectOptionId } from '@getodk/web-forms/lib/format/select-option-id.ts';
 import type { SelectNode } from '@getodk/xforms-engine';
 import Checkbox from 'primevue/checkbox';