+++ packages/web-forms/src/components/common/map/MapControls.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-import IconSVG from '@/components/common/IconSVG.vue';
-import ActionsInfoDialog, { type ActionInfo } from '@/components/common/ActionsInfoDialog.vue';
-import { TRANSLATE } from '@/lib/constants/injection-keys.ts';
-import type { Translate } from '@/lib/locale/useLocale.ts';
+import IconSVG from '@getodk/web-forms/components/common/IconSVG.vue';
+import ActionsInfoDialog, { type ActionInfo } from '@getodk/web-forms/components/common/ActionsInfoDialog.vue';
+import { TRANSLATE } from '@getodk/web-forms/lib/constants/injection-keys.ts';
+import type { Translate } from '@getodk/web-forms/lib/locale/useLocale.ts';
 import { computed, inject, ref } from 'vue';
 
 const props = defineProps<{