+++ packages/web-forms/src/components/form-elements/upload/UploadImageHeader.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-import { getModeConfig, MODES, type CanvasMode } from '@/components/common/canvas/getModeConfig.ts';
-import IconSVG from '@/components/common/IconSVG.vue';
-import { TRANSLATE } from '@/lib/constants/injection-keys.ts';
-import type { Translate } from '@/lib/locale/useLocale.ts';
+import { getModeConfig, MODES, type CanvasMode } from '@getodk/web-forms/components/common/canvas/getModeConfig.ts';
+import IconSVG from '@getodk/web-forms/components/common/IconSVG.vue';
+import { TRANSLATE } from '@getodk/web-forms/lib/constants/injection-keys.ts';
+import type { Translate } from '@getodk/web-forms/lib/locale/useLocale.ts';
 import type { UploadNode } from '@getodk/xforms-engine';
 import Button from 'primevue/button';
 import type { HTMLInputElementEvent, Ref } from 'vue';